+++ RELEASE_NOTES.rst
@@ -50,6 +50,7 @@ This is a maintenance release to catch up with *pytools* 1.1.4.
 - FIX: ensure correct weight labels when rendering dendrograms as plain text using the
   :class:`.DendrogramReportStyle`
 - FIX: calling method :meth:`.Id.get_class_id` could cause a :class:`.TypeError`
+- VIZ: set colors of axis labels to the foreground color of the current color scheme
 
 
 1.1.5